requirements.txt
django>=5.2.4
geopandas>=1.1.1
numpy>=2.3.1
google-ai-generativelanguage>=0.6.15
google-api-core>=2.25.1
google-api-python-client>=2.178.0
google-auth>=2.40.3
google-auth-httplib2>=0.2.0
google-cloud-aiplatform>=1.127.0
google-cloud-bigquery>=3.38.0
google-cloud-core>=2.5.0
google-cloud-resource-manager>=1.15.0
google-cloud-storage>=3.4.1
google-crc32c>=1.7.1
google-genai>=1.51.0
google-generativeai>=0.8.5
google-resumable-media>=2.8.0
googleapis-common-protos>=1.70.0
pandas>=2.3.0
pyproj>=3.7.1
shapely>=2.1.1
tqdm>=4.67.1